Fix some compile errors

+++ common/pipeline_storage/src/test/scala/uk/ac/wellcome/pipeline_storage/fixtures/ElasticIndexerFixtures.scala
@@ -22,7 +22,7 @@ case class SampleDocumentData(stuff: Option[String] = None)
 object SampleDocument {
 
   implicit val indexable = new Indexable[SampleDocument] {
-    def version(document: SampleDocument): Int = document.version
+    def version(document: SampleDocument): Long = document.version
     def id(document: SampleDocument): String = document.canonicalId
   }
